Fix how shmem keys are calculated so spproxy & proc agree

+++ cpp/shmem/shmem.cc
@@ -3,7 +3,7 @@
 namespace sigmaos {
 namespace shmem {
 
-uint64_t id2key(const std::string &key) {
+uint32_t id2key(const std::string &key) {
   // fnv64a hash inspired by
   // https://cs.opensource.google/go/go/+/refs/tags/go1.24.3:src/hash/fnv/fnv.go;l=65
   uint64_t s = 14695981039346656037LLU;
@@ -12,7 +12,7 @@ uint64_t id2key(const std::string &key) {
     s ^= (uint64_t)key[i];
     s *= prime64;
   }
-  return s;
+  return (uint32_t)s;
 }
 
 };  // namespace shmem
